fix: upgrade qulice-maven-plugin to 0.30.6 and fix all violations
Qulice 0.30.6 tightens checkstyle rules (code-free constructors,
JDK-obsolete APIs, ordered properties, OS-independent line separators,
among others). Every flagged violation across s3auth-hosts,
s3auth-relay and s3auth-rest got a genuine source fix, with no
suppressions added.

+++ s3auth-hosts/src/test/java/com/s3auth/hosts/H2DomainStatsDataTest.java
@@ -63,16 +63,15 @@ final class H2DomainStatsDataTest {
         );
         MatcherAssert.assertThat(
             stats.get(first).bytesTransferred(),
-            Matchers.is((long) 150)
+            Matchers.is(150L)
         );
         MatcherAssert.assertThat(
             stats.get(second).bytesTransferred(),
-            Matchers.is((long) 1000)
+            Matchers.is(1000L)
         );
         MatcherAssert.assertThat(
             data.all().size(),
             Matchers.is(0)
         );
     }
-
 }
